Refine gradle script (#1765)
