+++ core/build.gradle.kts
@@ -45,19 +45,7 @@ val apiCode: Int by rootProject.extra
 val verCode: Int by rootProject.extra
 val verName: String by rootProject.extra
 
-val androidTargetSdkVersion: Int by rootProject.extra
-val androidMinSdkVersion: Int by rootProject.extra
-val androidBuildToolsVersion: String by rootProject.extra
-val androidCompileSdkVersion: Int by rootProject.extra
-val androidCompileNdkVersion: String by rootProject.extra
-val androidSourceCompatibility: JavaVersion by rootProject.extra
-val androidTargetCompatibility: JavaVersion by rootProject.extra
-
 android {
-    compileSdk = androidCompileSdkVersion
-    ndkVersion = androidCompileNdkVersion
-    buildToolsVersion = androidBuildToolsVersion
-
     flavorDimensions += "api"
 
     buildFeatures {
@@ -66,41 +54,7 @@ android {
 
     defaultConfig {
         applicationId = "org.lsposed.lspd"
-        minSdk = androidMinSdkVersion
-        targetSdk = androidTargetSdkVersion
-        versionCode = verCode
-        versionName = verName
         multiDexEnabled = false
-
-        externalNativeBuild {
-            cmake {
-                arguments += "-DEXTERNAL_ROOT=${File(rootDir.absolutePath, "external")}"
-                abiFilters("arm64-v8a", "armeabi-v7a", "x86", "x86_64")
-                val flags = arrayOf(
-                    "-Wall",
-                    "-Qunused-arguments",
-                    "-Wno-gnu-string-literal-operator-template",
-                    "-fno-rtti",
-                    "-fvisibility=hidden",
-                    "-fvisibility-inlines-hidden",
-                    "-fno-exceptions",
-                    "-fno-stack-protector",
-                    "-fomit-frame-pointer",
-                    "-Wno-builtin-macro-redefined",
-                    "-Wno-unused-value",
-                    "-D__FILE__=__FILE_NAME__",
-                    "-DINJECTED_AID=$injectedPackageUid",
-                )
-                cppFlags("-std=c++20", *flags)
-                cFlags("-std=c18", *flags)
-                arguments(
-                    "-DANDROID_STL=none",
-                    "-DVERSION_CODE=$verCode",
-                    "-DVERSION_NAME=$verName",
-                )
-                targets("lspd")
-            }
-        }
 
         buildConfigField("int", "API_CODE", "$apiCode")
         buildConfigField(
@@ -112,58 +66,10 @@ android {
         buildConfigField("int", "MANAGER_INJECTED_UID", """$injectedPackageUid""")
     }
 
-    lint {
-        abortOnError = true
-        checkReleaseBuilds = false
-    }
-
     buildTypes {
-        debug {
-            externalNativeBuild {
-                cmake {
-                    arguments.addAll(
-                        arrayOf(
-                            "-DCMAKE_CXX_FLAGS_DEBUG=-Og",
-                            "-DCMAKE_C_FLAGS_DEBUG=-Og",
-                        )
-                    )
-                }
-            }
-        }
         release {
             isMinifyEnabled = true
             proguardFiles("proguard-rules.pro")
-
-            externalNativeBuild {
-                cmake {
-                    val flags = arrayOf(
-                        "-Wl,--exclude-libs,ALL",
-                        "-ffunction-sections",
-                        "-fdata-sections",
-                        "-Wl,--gc-sections",
-                        "-fno-unwind-tables",
-                        "-fno-asynchronous-unwind-tables",
-                        "-flto=thin",
-                        "-Wl,--thinlto-cache-policy,cache_size_bytes=300m",
-                        "-Wl,--thinlto-cache-dir=${buildDir.absolutePath}/.lto-cache",
-                    )
-                    cppFlags.addAll(flags)
-                    cFlags.addAll(flags)
-                    val configFlags = arrayOf(
-                        "-Oz",
-                        "-DNDEBUG"
-                    ).joinToString(" ")
-                    arguments.addAll(
-                        arrayOf(
-                            "-DCMAKE_CXX_FLAGS_RELEASE=$configFlags",
-                            "-DCMAKE_CXX_FLAGS_RELWITHDEBINFO=$configFlags",
-                            "-DCMAKE_C_FLAGS_RELEASE=$configFlags",
-                            "-DCMAKE_C_FLAGS_RELWITHDEBINFO=$configFlags",
-                            "-DDEBUG_SYMBOLS_PATH=${buildDir.absolutePath}/symbols",
-                        )
-                    )
-                }
-            }
         }
     }
 
@@ -171,11 +77,6 @@ android {
         cmake {
             path("src/main/jni/CMakeLists.txt")
         }
-    }
-
-    compileOptions {
-        targetCompatibility(androidTargetCompatibility)
-        sourceCompatibility(androidSourceCompatibility)
     }
 
     productFlavors {
@@ -215,10 +116,10 @@ dependencies {
     implementation("org.apache.commons:commons-lang3:3.12.0")
     implementation("de.upb.cs.swt:axml:2.1.2")
     compileOnly("androidx.annotation:annotation:1.3.0")
-    compileOnly(project(":hiddenapi:stubs"))
-    implementation(project(":hiddenapi:bridge"))
-    implementation(project(":manager-service"))
-    implementation(project(":daemon-service"))
+    compileOnly(projects.hiddenapi.stubs)
+    implementation(projects.hiddenapi.bridge)
+    implementation(projects.managerService)
+    implementation(projects.daemonService)
 }
 
 val zipAll = task("zipAll") {
