+++ hiddenapi/bridge/build.gradle.kts
@@ -22,5 +22,5 @@ plugins {
 }
 
 dependencies {
-    compileOnly(project(":hiddenapi:stubs"))
+    compileOnly(projects.hiddenapi.stubs)
 }
